// generated/src/app/api-models/site-address.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class SiteAddress {
  siteCode: string;
  seqSiteAddressId: number;
  termReasonCode: string;
  name1: string;
  name2: string;
  addressLine1: string;
  addressLine2: string;
  city: string;
  county: string;
  state: string;
  zipCode: string;
  sundayHours: string;
  mondayHours: string;
  tuesdayHours: string;
  wednesdayHours: string;
  thursdayHours: string;
  fridayHours: string;
  saturdayHours: string;
  mailingAddressLine1: string;
  mailingAddressLine2: string;
  mailingCity: string;
  mailingCounty: string;
  mailingState: string;
  mailingZipCode: string;
  siteaUserDefined1: string;
  siteaUserDefined2: string;
  siteaUserDate1: Date;
  siteaUserDate2: Date;
  insertProcess: string;
  insertUser: string;
  insertDatetime: Date;
  updateProcess: string;
  updateUser: string;
  updateDatetime: Date;
  securityCode: string;
  certifiedFlg: string;
  effectiveDate: Date;
  termDate: Date;
  primaryAddressFlg: string;
}